Final/requirements.txt
joblib
numpy
pandas
python-Levenshtein
scikit-learn
tensorflow
